requirements.txt
numpy==1.24.3
openai==0.27.8
pandas==1.5.3
python-dotenv==1.0.0
scikit_learn==1.2.2
spotipy==2.23.0
